fixup: rename to serde::displayfromstr, other nits.
